+++ src/main/java/com/anon/backend/entity/RefPostTag.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 * 
 *
 * @author anon
-* @since 2024-10-30
+* @since 2025-01-14
 */
 @Getter
 @Setter
@@ -25,11 +25,11 @@ public class RefPostTag implements Serializable {
   private static final long serialVersionUID = 1L;
 
   @TableId(value = "id", type = IdType.AUTO)
-  private Integer id;
+  private Long id;
 
-  private Integer postId;
+  private Long postId;
 
-  private Integer tagId;
+  private Long tagId;
 
   @TableLogic
   private Boolean isDelete;
